ENG-10087: Remove rules for scaling instances on azurerm_monitor_autoscale_setting.
We currently don't have these rules for EC2 sidecar due to the risk of scaling down and closing user connections. The scale set should have a fixed number of instances (minimum, maximum and desired) as we have in AWS EC2. Please see the parameters asg_* in the AWS EC2 module.